chore(release): package plugin, add ci, and bump to 1.2.0
- add GitHub Actions build+release packaging
- consolidate plugin into sp_usd_creator package with thin entrypoint
- update build/install scripts for root-level plugin layout
- add user/dev docs and update README install steps
- add tests and uv config; ignore build output and .idea
- bump VERSION to 1.2.0
