--- .idea/SP_usd_creator.iml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="PYTHON_MODULE" version="4">
-  <component name="NewModuleRootManager">
-    <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/scripts" isTestSource="false" />
-    </content>
-    <orderEntry type="jdk" jdkName="Python 3.9 virtualenv at G:\Projects\Dev\python_venvs\venv_001" jdkType="Python SDK" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-</module>
