--- .idea/misc.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="Black">
-    <option name="sdkName" value="Python 3.9 virtualenv at G:\Projects\Dev\python_venvs\venv_001" />
-  </component>
-</project>
