+++ README.md
@@ -15,10 +15,26 @@ This plugin allows you to export materials and geometry from Adobe Substance Pai
 
 
 ## Installation
-1. Copy `sp_plugin.py` and the `scripts` folder to your Substance Painter plugins directory, usually located at:
+1. Download the latest release zip from GitHub Releases (`AxeFX_usd_plugin.zip`).
+2. Unzip it into your Substance Painter plugins directory, usually located at:
    - Windows: `C:\Users\<USERNAME>\Documents\Adobe\Adobe Substance 3D Painter\python\plugins`
    - macOS: `~/Library/Application Support/Adobe/Adobe Substance 3D Painter/python/plugins`
    - Linux: `~/.local/share/Adobe/Adobe Substance 3D Painter/python/plugins`
+   This will place `AxeFX_usd_plugin.py` and `sp_usd_creator/` directly in the plugins folder.
+3. Restart Substance Painter.
+
+If you previously installed a folder named `AxeFX_usd_plugin/` or an older `sp_usd_creator/`, delete them to avoid duplicate plugins.
+
+## Developer Install
+1. Build the plugin bundle:
+   - `python tools/build_plugin.py`
+2. Copy `dist/AxeFX_usd_plugin.py` and `dist/sp_usd_creator/` into your Substance Painter plugins directory.
+3. Optional dev install on Windows:
+   - `powershell -File tools/install_plugin.ps1`
+
+## Documentation
+- User guide: `docs/USER_GUIDE.md`
+- Developer guide: `docs/DEVELOPER_GUIDE.md`
 
 
 ## Usage
@@ -27,7 +43,7 @@ This plugin allows you to export materials and geometry from Adobe Substance Pai
    - If the plugin is not visible, ensure it is installed in the correct directory.
 3. Configure the export settings:
    - **Render Engines**: Select the render engines you want to export (USD Preview, Arnold, MaterialX).
-   - **Publish Location**: Set the path where the USD file will be saved (e.g., `<export_folder>/materials.usda`).
+   - **Publish Directory**: Set the directory where USD files will be created (e.g., `<export_folder>`).
    - **Primitive Path**: Set the path for the root node in the USD file (e.g., `/RootNode/Materials`).
    - **Save Geometry**: Check this option to export mesh geometry as a separate USD file (e.g., `mesh.usd`).
 4. Export the textures normally in Substance Painter. The plugin will run automatically after the export process.
@@ -43,7 +59,7 @@ This plugin allows you to export materials and geometry from Adobe Substance Pai
 
 1. Open the plugin UI and configure the following settings:
    - **Render Engines**: Enable `USD Preview` and `Arnold`.
-   - **Publish Location**: Set to `<export_folder>/materials.usda`.
+   - **Publish Directory**: Set to `<export_folder>`.
    - **Primitive Path**: Set to `/RootNode/Materials`.
    - **Save Geometry**: Check this option to export mesh geometry.
