+++ pyproject.toml
@@ -1,0 +1,13 @@
+[project]
+name = "sp-usd-creator"
+version = "0.0.0"
+requires-python = ">=3.9"
+dependencies = [
+  "usd-core",
+]
+
+[dependency-groups]
+dev = [
+  "pytest",
+  "ruff",
+]
