+++ scripts/core/models.py
@@ -1,0 +1,1 @@
+from sp_usd_creator.core.models import *
