+++ scripts/core/publish_paths.py
@@ -1,0 +1,1 @@
+from sp_usd_creator.core.publish_paths import *
