+++ scripts/core/texture_parser.py
@@ -1,0 +1,1 @@
+from sp_usd_creator.core.texture_parser import *
